backend/requirements.txt
django==3.*
djangorestframework
psycopg2-binary